requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
openai==1.12.0
pyyaml==6.0.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-dotenv==1.0.1
